requirements.txt
numpy
pandas
pytorch
onos-ric-sdk-python>=0.2.5